% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{compRelSEM}
\alias{compRelSEM}
\title{Composite Reliability using SEM}
\usage{
compRelSEM(object, obs.var = TRUE, tau.eq = FALSE, ord.scale = TRUE,
  return.total = FALSE, dropSingle = TRUE, omit.factors = character(0),
  omit.indicators = character(0), omit.imps = c("no.conv", "no.se"))
}
\arguments{
\item{object}{A \code{\linkS4class{lavaan}} or
\code{\linkS4class{lavaan.mi}} object, expected to contain only
exogenous common factors (i.e., a CFA model).}

\item{obs.var}{\code{logical} indicating whether to compute AVE using
observed variances in the denominator. Setting \code{FALSE} triggers
using model-implied variances in the denominator.}

\item{tau.eq}{\code{logical} indicating whether to assume (essential)
tau-equivalence, yielding a coefficient analogous to \eqn{\alpha}.
Setting \code{FALSE} yields an \eqn{\omega}-type coefficient.}

\item{ord.scale}{\code{logical} indicating whether to apply Green and Yang's
(2009, formula 21) correction, so that reliability is calculated for the
actual ordinal response scale (ignored for factors with continuous
indicators).  Setting \code{FALSE} yields coefficients that are
only applicable to the continuous latent-response scale.}

\item{return.total}{\code{logical} indicating whether to return a final
column containing the reliability of a composite of all indicators (not
listed in \code{omit.indicators}) of factors not listed in
\code{omit.factors}.  Ignored in 1-factor models, and should only be set
\code{TRUE} if all factors represent scale dimensions that could be
meaningfully collapsed to a single composite (scale sum or scale mean).
Setting a negative value (e.g., \code{-1} returns **only** the
total-composite reliability (excluding coefficients per factor).}

\item{dropSingle}{\code{logical} indicating whether to exclude factors
defined by a single indicator from the returned results. If \code{TRUE}
(default), single indicators will still be included in the \code{total}
column when \code{return.total = TRUE}.}

\item{omit.factors}{\code{character} vector naming any common factors
modeled in \code{object} whose composite reliability is not of
interest. For example, higher-order or method factors. Note that
\code{\link{reliabilityL2}()} should be used to calculate composite
reliability of a higher-order factor.}

\item{omit.indicators}{\code{character} vector naming any observed variables
that should be ignored when calculating composite reliability. This can
be useful, for example, to estimate reliability when an indicator is
removed.}

\item{omit.imps}{\code{character} vector specifying criteria for omitting
imputations from pooled results.  Can include any of
\code{c("no.conv", "no.se", "no.npd")}, the first 2 of which are the
default setting, which excludes any imputations that did not
converge or for which standard errors could not be computed.  The
last option (\code{"no.npd"}) would exclude any imputations which
yielded a nonpositive definite covariance matrix for observed or
latent variables, which would include any "improper solutions" such
as Heywood cases.  NPD solutions are not excluded by default because
they are likely to occur due to sampling error, especially in small
samples.  However, gross model misspecification could also cause
NPD solutions, users can compare pooled results with and without
this setting as a sensitivity analysis to see whether some
imputations warrant further investigation.}
}
\value{
A \code{numeric} vector of composite reliability coefficients per
  factor, or a \code{list} of vectors per "block" (group and/or level of
  analysis). If there are multiple factors, whose multidimensional
  indicators combine into a single composite, users can request
  \code{return.total=TRUE} to add a column including a reliability
  coefficient for the total composite, or.\code{return.total = -1} to
  return **only** the total-composite reliability.
}
\description{
Calculate composite reliability from estimated factor-model parameters
}
\details{
Several coefficients for factor-analysis reliability have been termed
"omega", which Cho (2021) argues is a misleading misnomer and argues for
using \eqn{\rho} to represent them all, differentiated by descriptive
subscripts.  In our package, we number \eqn{\omega} based on commonly
applied calculations.

Bentler (1968) first introduced factor-analysis reliability for a
unidimensional factor model with congeneric indicators, labeling the
coeficients \eqn{\alpha}.  McDonald (1999) later referred to this
\emph{and other reliability coefficients}, first as \eqn{\theta} (in 1970),
then as \eqn{\omega}, which is a source of confusion when reporting
coefficients (Cho, 2021).  Coefficients based on factor models were later
generalized to account for multidimenisionality (possibly with
cross-loadings) and correlated errors. The general \eqn{\omega} formula
implemented in this function is:

\deqn{ \omega = \frac{\left( \sum^{k}_{i = 1} \lambda_i \right)^{2}
Var\left( \psi \right)}{\bold{1}^\prime \hat{\Sigma} \bold{1}}, }

where \eqn{\hat{\Sigma}} can be the model-implied covariance matrix from
either the saturated model (i.e., the "observed" covariance matrix, used by
default) or from the hypothesized CFA model, controlled by the
\code{obs.var} argument. A \eqn{k}-dimensional vector \eqn{\bold{1}} is used
to sum elements in the matrix. Note that if the model includes any directed
effects (latent regression slopes), all coefficients are calculated
from **total** factor variances:
\code{\link[lavaan]{lavInspect}(object, "cov.lv")}.

Assuming (essential) tau-equivalence makes \eqn{\omega} equivalent to the
coefficient \eqn{\alpha} from classical test theory (Cronbach, 1951):

\deqn{ \alpha = \frac{k}{k - 1}\left[ 1 - \frac{\sum^{k}_{i = 1}
\sigma_{ii}}{\sum^{k}_{i = 1} \sigma_{ii} + 2\sum_{i < j} \sigma_{ij}}
\right],}

where \eqn{k} is the number of items in a factor's compositte,
\eqn{\sigma_{ii}} signifies item \emph{i}'s variance, and \eqn{\sigma_{ij}}
signifies the covariance between items \emph{i} and \emph{j}. Again, the
\code{obs.var} argument controls whether \eqn{\alpha} is calculated using
the observed or model-implied covariance matrix.

By setting \code{return.total=TRUE}, one can estimate reliability for a
single composite calculated using all indicators in a multidimensional
CFA (Bentler, 1972, 2009). Setting \code{return.total = -1} will return
**only** the total-composite reliability (not per factor).

When all indicators (per composite) are ordinal, the \code{ord.scale}
argument controls whether the coefficient is calculated on the
latent-response scale (\code{FALSE}) or on the observed ordinal scale
(\code{TRUE}, the default).  For \eqn{\omega}-type coefficients
(\code{tau.eq=FALSE}), Green and Yang's (2009, formula 21) approach is used
to transform factor-model results back to the ordinal response scale.
When \code{ord.scale=TRUE}, coefficient \eqn{\alpha} is calculated using the
covariance matrix calculated from the integer-valued numeric weights for
ordinal categories, consistent with its definition (Chalmers, 2018) and the
\code{alpha} function in the \code{psych} package; this implies
\code{obs.var=TRUE}, so \code{obs.var=FALSE} will be ignored.  When
\code{ord.scale=FALSE}, the standard \eqn{\alpha} formula is applied to the
polychoric correlation matrix ("ordinal \eqn{\alpha}"; Zumbo et al., 2007),
estimated from the saturated or hypothesized model (see \code{obs.var}),
and \eqn{\omega} is calculated from CFA results without applying Green and
Yang's (2009) correction (see Zumbo & Kroc's, 2019, for a rationalization).
No method has been proposed for calculating reliability with a mixture of
categorical and continuous indicators, so an error is returned if
\code{object} includes factors with a mixture of indicator types (unless
omitted using \code{omit.factors}). If categorical indicators load on a
different factor(s) than continuous indicators, then reliability will still
be calculated separately for those factors, but \code{return.total} must be
\code{FALSE} (unless \code{omit.factors} is used to isolate factors with
indicators of the same type).
}
\examples{

data(HolzingerSwineford1939)
HS9 <- HolzingerSwineford1939[ , c("x7","x8","x9")]
HSbinary <- as.data.frame( lapply(HS9, cut, 2, labels=FALSE) )
names(HSbinary) <- c("y7","y8","y9")
HS <- cbind(HolzingerSwineford1939, HSbinary)

HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ y7 + y8 + y9 '

fit <- cfa(HS.model, data = HS, ordered = c("y7","y8","y9"), std.lv = TRUE)

## works for factors with exclusively continuous OR categorical indicators
compRelSEM(fit)

## reliability for ALL indicators only available when they are
## all continuous or all categorical
compRelSEM(fit, omit.factors = "speed", return.total = TRUE)


## loop over visual indicators to calculate alpha if one indicator is removed
for (i in paste0("x", 1:3)) {
  cat("Drop x", i, ":\n")
  print(compRelSEM(fit, omit.factors = c("textual","speed"),
                   omit.indicators = i, tau.eq = TRUE))
}


## works for multigroup models and for multilevel models (and both)
data(Demo.twolevel)
## assign clusters to arbitrary groups
Demo.twolevel$g <- ifelse(Demo.twolevel$cluster \%\% 2L, "type1", "type2")
model2 <- ' group: type1
  level: within
    fac =~ y1 + L2*y2 + L3*y3
  level: between
    fac =~ y1 + L2*y2 + L3*y3

group: type2
  level: within
    fac =~ y1 + L2*y2 + L3*y3
  level: between
    fac =~ y1 + L2*y2 + L3*y3
'
fit2 <- sem(model2, data = Demo.twolevel, cluster = "cluster", group = "g")
compRelSEM(fit2)



}
\references{
Bentler, P. M. (1972). A lower-bound method for the dimension-free
measurement of internal consistency. \emph{Social Science Research, 1}(4),
343--357. \doi{10.1016/0049-089X(72)90082-8}

Bentler, P. M. (2009). Alpha, dimension-free, and model-based internal
consistency reliability. \emph{Psychometrika, 74}(1), 137--143.
\doi{10.1007/s11336-008-9100-1}

Chalmers, R. P. (2018). On misconceptions and the limited usefulness of
ordinal alpha. \emph{Educational and Psychological Measurement, 78}(6),
1056--1071. \doi{10.1177/0013164417727036}

Cho, E. (2021) Neither Cronbach’s alpha nor McDonald’s omega: A commentary
on Sijtsma and Pfadt. *Psychometrika, 86*(4), 877--886.
\doi{10.1007/s11336-021-09801-1}

Cronbach, L. J. (1951). Coefficient alpha and the internal structure of
tests. \emph{Psychometrika, 16}(3), 297--334. \doi{10.1007/BF02310555}

Green, S. B., & Yang, Y. (2009). Reliability of summed item scores using
structural equation modeling: An alternative to coefficient alpha.
\emph{Psychometrika, 74}(1), 155--167. \doi{10.1007/s11336-008-9099-3}

McDonald, R. P. (1999). \emph{Test theory: A unified treatment}. Mahwah, NJ:
Erlbaum.

Zumbo, B. D., Gadermann, A. M., & Zeisser, C. (2007). Ordinal versions of
coefficients alpha and theta for Likert rating scales.
\emph{Journal of Modern Applied Statistical Methods, 6}(1), 21--29.
\doi{10.22237/jmasm/1177992180}

Zumbo, B. D., & Kroc, E. (2019). A measurement is a choice and Stevens’
scales of measurement do not help make it: A response to Chalmers.
\emph{Educational and Psychological Measurement, 79}(6), 1184--1197.
\doi{10.1177/0013164419844305}
}
\seealso{
\code{\link{reliabilityL2}} for reliability value of a desired
second-order factor, \code{\link{maximalRelia}} for the maximal reliability
of weighted composite
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})

  Uses hidden functions written by Sunthud Pornprasertmanit
  (\email{psunthud@gmail.com}) for original \code{reliability()} function.
}
