% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat2way}
\alias{dat2way}
\title{Simulated Dataset to Demonstrate Two-way Latent Interaction}
\format{
A \code{data.frame} with 500 observations of 9 variables.
\describe{
\item{x1}{The first indicator of the first independent factor}
\item{x2}{The second indicator of the first independent factor}
\item{x3}{The third indicator of the first independent factor}
\item{x4}{The first indicator of the second independent factor}
\item{x5}{The second indicator of the second independent factor}
\item{x6}{The third indicator of the second independent factor}
\item{x7}{The first indicator of the dependent factor}
\item{x8}{The second indicator of the dependent factor}
\item{x9}{The third indicator of the dependent factor}
}
}
\source{
Data were generated by the \code{\link[MASS]{mvrnorm}} function in
  the \code{MASS} package.
}
\usage{
dat2way
}
\description{
A simulated data set with 2 independent factors and 1 dependent factor where
each factor has three indicators
}
\examples{
 head(dat2way)
}
\keyword{datasets}
