\name{dat3way}
\alias{dat3way}
\title{
Simulated Dataset to Demonstrate Three-way Latent Interaction
}
\description{
A simulated data set with 3 independent factors and 1 dependent factor where each factor has three indicators 
}
\usage{
data(dat3way)
}
\format{
  A data frame with 500 observations of 12 variables.
	\describe{
   \item{x1}{The first indicator of the first independent factor}
   \item{x2}{The second indicator of the first independent factor}
   \item{x3}{The third indicator of the first independent factor}
   \item{x4}{The first indicator of the second independent factor}
   \item{x5}{The second indicator of the second independent factor}
   \item{x6}{The third indicator of the second independent factor}
   \item{x7}{The first indicator of the third independent factor}
   \item{x8}{The second indicator of the third independent factor}
   \item{x9}{The third indicator of the third independent factor}
   \item{x10}{The first indicator of the dependent factor}
   \item{x11}{The second indicator of the dependent factor}
   \item{x12}{The third indicator of the dependent factor}
   }
}
\source{
Data was generated by the \link[MASS]{mvrnorm} function in the \code{MASS} package.
}
\examples{
head(dat3way)
}
