% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_lavaan.R
\name{fitted_lavaan}
\alias{fitted_lavaan}
\title{Fitted Values from Observed Variable Model SEM}
\usage{
fitted_lavaan(fit)
}
\arguments{
\item{fit}{A \code{lavaan} object fit using observed variables only and 
\code{meanstructure=TRUE}.}
}
\value{
This function returns a \code{data.frame} with the fitted values
of endogenous variables.
}
\description{
Implements fitted values for observed variable models fit by \code{lavaan}.
Does not yet work for latent variable models.
}
\examples{


iris_mod <- '
Sepal.Width ~ Sepal.Length
Petal.Width ~ Sepal.Width + Petal.Length
'
iris_fit <- sem(iris_mod, data=iris, meanstructure = TRUE)

fitted_lavaan(iris_fit)                  


}
\author{
Jarrett E.K. Byrnes, University of Massachusetts Boston
\email{jarrett.byrnes@umb.edu})
}
\seealso{
\code{\link{predict_lavaan}}, \code{\link{residuals_lavaan}}, code{\link[lavaan]{lavOptions}}
}

