% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMI-score.R
\name{lavTestScore.mi}
\alias{lavTestScore.mi}
\title{Score Test for Multiple Imputations}
\usage{
lavTestScore.mi(object, add = NULL, release = NULL, type = c("D2",
  "Rubin"), asymptotic = FALSE, univariate = TRUE, cumulative = FALSE,
  epc = FALSE, verbose = FALSE, warn = TRUE)
}
\arguments{
\item{type}{\code{character} indicating which pooling method to use.
\code{"Rubin"} indicates Rubin's (1987) rules will be applied to the
gradient and information, and those pooled values will be used to
calculate modification indices in the usual manner. \code{"D2"} (default),
\code{"LMRR"}, or \code{"Li.et.al"} indicate that modification indices
calculated from each imputed data set will be pooled across imputations,
as described in Li, Meng, Raghunathan, & Rubin (1991) and Enders (2010).}

\item{asymptotic}{\code{logical}. If \code{FALSE}, the pooled test will be
returned as an \emph{F}-distributed variable with numerator (\code{df1})
and denominator (\code{df2}) degrees of freedom.  If \code{TRUE} (default),
the pooled \emph{F} statistic will be multiplied by its \code{df1} on the
assumption that its \code{df2} is sufficiently large enough that the
statistic will be asymptotically \eqn{\chi^2} distributed with \code{df1}.}
}
\value{
A list containing at least one \code{data.frame} (the total score
test), and optionally containing each 1-\emph{df} score test, the cumulative
score tests, and expected parameter changes.
See \code{\link[lavaan]{lavTestScore}} for details.
}
\description{
Score test (or Lagrange Multiplier test) for lavaan models fitted to
multiple imputed data sets. Statistics for releasing one or more
fixed or constrained parameters in model can be calculated by pooling
the gradient and information matrices pooled across imputed data sets
using Rubin's (1987) rules, or by pooling the score test statistics
across imputed data sets (Li, Meng, Raghunathan, & Rubin, 1991).
}
\examples{
 \dontrun{
## impose missing data for example
HSMiss <- HolzingerSwineford1939[ , c(paste("x", 1:9, sep = ""),
                                      "ageyr","agemo","school")]
set.seed(12345)
HSMiss$x5 <- ifelse(HSMiss$x5 <= quantile(HSMiss$x5, .3), NA, HSMiss$x5)
age <- HSMiss$ageyr + HSMiss$agemo/12
HSMiss$x9 <- ifelse(age <= quantile(age, .3), NA, HSMiss$x9)

## impute missing data
library(Amelia)
set.seed(12345)
HS.amelia <- amelia(HSMiss, m = 20, noms = "school", p2s = FALSE)
imps <- HS.amelia$imputations

## specify CFA model from lavaan's ?cfa help page
HS.model <- '
  speed =~ c(L1, L1)*x7 + c(L1, L1)*x8 + c(L1, L1)*x9
'

out <- cfa.mi(HS.model, data = imps, group = "school", std.lv = TRUE)

## Mode 1: Score test for releasing equality constraints

## default: Li et al.'s (1991) "D2" method
lavTestScore.mi(fit0, cumulative = TRUE)
lavTestScore.mi(out, type = "Rubin") # Rubin's rules

## Mode 2: Score test for adding currently fixed-to-zero parameters
lavTestScore.mi(fit0, add = 'x7 ~~ x8 + x9')

}

}
\references{
Bentler, P. M., & Chou, C.-P. (1992). Some new covariance structure model
improvement statistics. \emph{Sociological Methods & Research, 21}(2),
259--282. doi:10.1177/0049124192021002006

Enders, C. K. (2010). \emph{Applied missing data analysis}.
New York, NY: Guilford.

Li, K.-H., Meng, X.-L., Raghunathan, T. E., & Rubin, D. B. (1991).
Significance levels from repeated \emph{p}-values with multiply-imputed data.
\emph{Statistica Sinica, 1}(1), 65--92. Retrieved from
\url{http://www.jstor.org/stable/24303994}

Rubin, D. B. (1987). \emph{Multiple imputation for nonresponse in surveys}.
New York, NY: Wiley.
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})

Adapted from \pkg{lavaan} source code, written by
  Yves Rosseel (Ghent University; \email{Yves.Rosseel@UGent.be})

\code{type = "Rubin"} method proposed by
  Maxwell Mansolf (University of California, Los Angeles;
  \email{mamansolf@gmail.com})
}
