% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMI-modification.R
\name{modindices.mi}
\alias{modindices.mi}
\alias{modificationIndices.mi}
\alias{modificationindices.mi}
\title{Modification Indices for Multiple Imputations}
\usage{
modindices.mi(object, type = c("D2", "Rubin"), standardized = TRUE,
  cov.std = TRUE, power = FALSE, delta = 0.1, alpha = 0.05,
  high.power = 0.75, sort. = FALSE, minimum.value = 0,
  maximum.number = nrow(LIST), na.remove = TRUE, op = NULL)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{lavaan.mi}}}

\item{type}{\code{character} indicating which pooling method to use.
\code{"Rubin"} indicates Rubin's (1987) rules will be applied to the
gradient and information, and those pooled values will be used to
calculate modification indices in the usual manner. \code{"D2"} (default),
\code{"LMRR"}, or \code{"Li.et.al"} indicate that modification indices
calculated from each imputed data set will be pooled across imputations,
as described in Li, Meng, Raghunathan, & Rubin (1991) and Enders (2010).}

\item{cov.std}{\code{logical}. Ignored if \code{type == "D2"}.
  If \code{TRUE} (default), the (residual)
observed covariances are scaled by the square-root of the diagonal elements
of the \eqn{\Theta} matrix, and the (residual) latent covariances are
scaled by the square-root of the diagonal elements of the \eqn{\Psi}
matrix. If \code{FALSE}, the (residual) observed covariances are scaled by
the square-root of the diagonal elements of the model-implied covariance
matrix of observed variables (\eqn{\Sigma}), and the (residual) latent
covariances are scaled by the square-root of the diagonal elements of the
model-implied covariance matrix of the latent variables.}
}
\value{
A \code{data.frame} containing modification indices and (S)EPCs.
}
\description{
Modification indices (1-\emph{df} Lagrange Multiplier tests) from a
latent variable model fitted to multiple imputed data sets. Statistics
for releasing one or more fixed or constrained parameters in model can
be calculated by pooling the gradient and information matrices pooled
across imputed data sets using Rubin's (1987) rules, or by pooling the
modification indices across imputed data sets (Li, Meng, Raghunathan, &
Rubin, 1991).
}
\note{
When \code{type = "D2"}, (S)EPCs will be pooled by taking the average
across imputations. When \code{type = "Rubin"}, EPCs will be calculated
in the standard way using the pooled gradient and information, and SEPCs
will be calculated by standardizing the EPCs using model-implied (residual)
variances.
}
\examples{
 \dontrun{
## impose missing data for example
HSMiss <- HolzingerSwineford1939[ , c(paste("x", 1:9, sep = ""),
                                      "ageyr","agemo","school")]
set.seed(12345)
HSMiss$x5 <- ifelse(HSMiss$x5 <= quantile(HSMiss$x5, .3), NA, HSMiss$x5)
age <- HSMiss$ageyr + HSMiss$agemo/12
HSMiss$x9 <- ifelse(age <= quantile(age, .3), NA, HSMiss$x9)

## impute missing data
library(Amelia)
set.seed(12345)
HS.amelia <- amelia(HSMiss, m = 20, noms = "school", p2s = FALSE)
imps <- HS.amelia$imputations

## specify CFA model from lavaan's ?cfa help page
HS.model <- '
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
'

out <- cfa.mi(HS.model, data = imps)

modindices.mi(out) # default: Li et al.'s (1991) "D2" method
modindices.mi(out, type = "Rubin") # Rubin's rules

}

}
\references{
Enders, C. K. (2010). \emph{Applied missing data analysis}.
New York, NY: Guilford.

Li, K.-H., Meng, X.-L., Raghunathan, T. E., & Rubin, D. B. (1991).
Significance levels from repeated \emph{p}-values with multiply-imputed data.
\emph{Statistica Sinica, 1}(1), 65--92. Retrieved from
\url{http://www.jstor.org/stable/24303994}

Rubin, D. B. (1987). \emph{Multiple imputation for nonresponse in surveys}.
New York, NY: Wiley.
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})

  Yves Rosseel (Ghent University; \email{Yves.Rosseel@UGent.be})

  Maxwell Mansolf (University of California, Los Angeles;
  \email{mamansolf@gmail.com})
}
