% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET.R
\docType{class}
\name{Net-class}
\alias{Net-class}
\alias{show,Net-method}
\alias{summary,Net-method}
\alias{show,Net-method}
\alias{summary,Net-method}
\title{Class For the Result of Nesting and Equivalence Testing}
\usage{
\S4method{show}{Net}(object)

\S4method{summary}{Net}(object)
}
\value{
The \code{show} method displays the logical matrix of test results,
 and the \code{summary} method is used to describe the results in narrative.
 In either case, the original \code{object} is invisibly returned.
}
\description{
This class contains the results of nesting and equivalence testing among
multiple models
}
\section{Slots}{

\describe{
\item{\code{test}}{Logical \code{matrix} indicating nesting/equivalence among models}

\item{\code{df}}{The degrees of freedom of tested models}
}}

\section{Objects from the Class}{
 Objects can be created via the
\code{\link{net}} function.
}

\examples{

# See the example in the net function.

}
\seealso{
\code{\link{net}}
}
\author{
Terrence D. Jorgensen (University of Amsterdam;
\email{TJorgensen314@gmail.com})
}
