% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMI.R
\name{runMI-deprecated}
\alias{runMI-deprecated}
\alias{lavaan.mi-deprecated}
\alias{cfa.mi-deprecated}
\alias{sem.mi-deprecated}
\alias{growth.mi-deprecated}
\title{Fit a lavaan Model to Multiple Imputed Data Sets}
\usage{
runMI(...)

lavaan.mi(...)

cfa.mi(...)

sem.mi...)

growth.mi(...)
}
\arguments{
\item{\dots}{Functionality has ceased to be supported, replaced by \code{lavaan.mi} package}
}
\value{
An error message
}
\description{
This functionality has been moved to the \code{lavaan.mi} package.
}
\examples{

## See the new lavaan.mi package

}
\seealso{
\code{\link[=semTools-deprecated]{semTools-deprecated()}}

\code{\link[=semTools-deprecated]{semTools-deprecated()}}

\code{\link[=semTools-deprecated]{semTools-deprecated()}}

\code{\link[=semTools-deprecated]{semTools-deprecated()}}

\code{\link[=semTools-deprecated]{semTools-deprecated()}}
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})
}
\keyword{internal}
