% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semTools-deprecated.R, R/standardizeMx.R
\name{semTools-deprecated}
\alias{semTools-deprecated}
\alias{standardizeMx}
\title{Deprecated functions in package \pkg{semTools}.}
\usage{
standardizeMx(object, free = TRUE)
}
\description{
The functions listed below are deprecated and will be defunct
  in the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("semTools-deprecated")}.
}
\section{\code{standardizeMx}}{

The \code{standardizeMx} and \code{fitMeasuresMx} functions will no longer
be supported, nor will there be replacement functions.  Their functionality
is now available in the \pkg{OpenMx} package, making these functions
obsolete. The utility functions \code{nullMx} and \code{saturateMx} will
also no longer be supported. These have already been removed from
\pkg{semTools}, except that \code{standardizeMx} remains deprecated due to
the temporary depndency on it of the \pkg{semPlot} package.  The exception
is that \code{\link[OpenMx]{mxStandardizeRAMpaths}} currently only provides
standardized estimates of covariance-structure parameters, whereas
\code{standardizeMx} also provides standardized means.
}

\keyword{internal}
