\name{simParcel}
\alias{simParcel}
\title{
Simulated Data set to Demonstrate Random Allocations of Parcels
}

\description{
A simulated data set with 3 factors with 9 indicators for each factor
}
\usage{
data(simParcel)
}
\format{
   \item{f1item1} {Item 1 loading on factor 1}
   \item{f1item2} {Item 2 loading on factor 1}
   \item{f1item3} {Item 3 loading on factor 1}
   \item{f2item1} {Item 1 loading on factor 2}
   \item{f2item2} {Item 2 loading on factor 2}
   \item{f2item3} {Item 3 loading on factor 2}
   \item{f3item1} {Item 1 loading on factor 3}
   \item{f3item2} {Item 2 loading on factor 3}
   \item{f3item3} {Item 3 loading on factor 3}
   }
 
\source{
Data was generated using the {\link{simsem}} package.
}
\examples{
head(simParcel)
}