% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{efaUnrotate}
\alias{efaUnrotate}
\title{Analyze Unrotated Exploratory Factor Analysis Model}
\usage{
efaUnrotate(data = NULL, nf, varList = NULL, start = TRUE, aux = NULL, ...)
}
\arguments{
\item{data}{A target \code{data.frame}}

\item{nf}{The desired number of factors}

\item{varList}{Target observed variables. If not specified, all variables in
\code{data} will be used (or \code{sample.cov} if \code{is.null(data)};
see \code{\link[lavaan]{cfa}} for argument descriptions).}

\item{start}{Use starting values in the analysis from the
\code{\link{factanal}} \code{function}. If \code{FALSE}, the starting values
from the \code{lavaan} package will be used. \code{TRUE} is ignored with a
warning if the \code{aux} argument is used.}

\item{aux}{The list of auxiliary variables. These variables will be included
in the model by the saturated-correlates approach to account for missing
information.}

\item{\dots}{Other arguments in the \code{\link[lavaan]{cfa}} function in
the \code{lavaan} package, such as \code{ordered}, \code{se},
\code{estimator}, or \code{sample.cov} and \code{sample.nobs}.}
}
\value{
A \code{lavaan} output of unrotated exploratory factor analysis
solution.
}
\description{
This function will analyze unrotated exploratory factor analysis model. The
unrotated solution can be rotated by the \code{\link{orthRotate}} and
\code{\link{oblqRotate}} functions.
}
\details{
This function will generate a lavaan script for unrotated exploratory factor
analysis model such that (1) all factor loadings are estimated, (2) factor
variances are fixed to 1, (3) factor covariances are fixed to 0, and (4) the
dot products of any pairs of columns in the factor loading matrix are fixed
to zero (Johnson & Wichern, 2002). The reason for creating this function
to supplement the \code{\link{factanal}} function is that users can enjoy
some advanced features from the \code{lavaan} package, such as scaled
\eqn{\chi^2}, diagonally weighted least squares estimation for ordinal
indicators, or full-information maximum likelihood (FIML) to handle
incomplete data.
}
\examples{

unrotated <- efaUnrotate(HolzingerSwineford1939, nf = 3,
                         varList = paste0("x", 1:9), estimator = "mlr")
summary(unrotated, std = TRUE)
inspect(unrotated, "std")

dat <- data.frame(HolzingerSwineford1939,
                  z = rnorm(nrow(HolzingerSwineford1939), 0, 1))
unrotated2 <- efaUnrotate(dat, nf = 2, varList = paste0("x", 1:9), aux = "z")

}
\author{
Sunthud Pornprasertmanit (\email{psunthud@gmail.com})

Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})
}
