% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET.R
\docType{class}
\name{Net-class}
\alias{Net-class}
\alias{show,Net-method}
\alias{summary,Net-method}
\title{Class For the Result of Nesting and Equivalence Testing}
\usage{
\S4method{show}{Net}(object)

\S4method{summary}{Net}(object)
}
\arguments{
\item{object}{An object of class \code{Net}.}
}
\value{
\item{show}{\code{signature(object = "Net")}: prints the logical matrix of
  test results. \code{NA} indicates a model did not converge.}
\item{summary}{\code{signature(object = "Net")}: prints a narrative
  description of results. The original \code{object} is invisibly returned.}
}
\description{
This class contains the results of nesting and equivalence testing among
multiple models
}
\section{Slots}{

\describe{
\item{\code{test}}{Logical \code{matrix} indicating nesting/equivalence among models}

\item{\code{df}}{The degrees of freedom of tested models}
}}

\section{Objects from the Class}{
 Objects can be created via the
\code{\link{net}} function.
}

\examples{

# See the example in the net function.

}
\seealso{
\code{\link{net}}
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})
}
