% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_semTools-deprecated.R, R/efa.R,
%   R/reliability.R, R/runMI.R
\name{semTools-deprecated}
\alias{semTools-deprecated}
\alias{efaUnrotate}
\alias{orthRotate}
\alias{oblqRotate}
\alias{funRotate}
\alias{reliability}
\alias{reliabilityL2}
\alias{runMI}
\alias{lavaan.mi}
\alias{cfa.mi}
\alias{sem.mi}
\alias{growth.mi}
\title{Deprecated functions in package \pkg{semTools}.}
\usage{
efaUnrotate(data = NULL, nf, varList = NULL, start = TRUE, aux = NULL,
  ...)

orthRotate(object, method = "varimax", ...)

oblqRotate(object, method = "quartimin", ...)

funRotate(object, fun, ...)

reliability(object, what = c("alpha", "omega", "omega2", "omega3", "ave"),
  return.total = FALSE, dropSingle = TRUE, omit.factors = character(0),
  omit.indicators = character(0), omit.imps = c("no.conv", "no.se"))

reliabilityL2(object, secondFactor, omit.imps = c("no.conv", "no.se"))

runMI(...)

lavaan.mi(...)

cfa.mi(...)

sem.mi(...)

growth.mi(...)
}
\description{
The functions listed below are deprecated and will be defunct
in the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("semTools-deprecated")}.
}
\section{Support functions for exploratory factor analysis in \code{lavaan}}{

The \code{efaUnrotate}, \code{orthRotate}, \code{oblqRotate}, and
\code{funRotate} functions will no longer be supported. These functions
allowed users to estimate EFA parameters with \code{lavaan}. Instead, users
can now directly use \code{lavaan}'s \code{\link[lavaan]{efa}()} function.
Exploratory SEM (ESEM) is also supported by lavaan using special operators
in \code{lavaan}'s \code{\link[lavaan]{model.syntax}}; see
\url{https://github.com/yrosseel/lavaan/issues/112} for details.
}

\section{Reliability}{

The original \code{reliability} function was suboptimally designed.
For example, AVE was returned, which is not a reliability index. Also,
alpha and several omega-type coefficients were returned, including the
original formula that was inappropriate for models with complex structure.
Some features could be controlled by the user for one but not both types of
index  For example, alpha for categorical indicators was returned on both
the observed and latent-response scales, but this was not an option for any
omega-type indices.  The omegas differed in terms of whether the observed or
model-implied covariance matrix was used in the denominator, but alpha was
only computed using the observed matrix.  These inconsistencies have been
resolved in the new \code{\link[=compRelSEM]{compRelSEM()}} function, which returns only
one reliability index per composite, tailored to the user's requested
features, for which there is much more flexibility. The average variance
extracted is now available in a dedicated \code{\link[=AVE]{AVE()}} function.
}

\section{Higher-Order Reliability}{

Originally, composite reliability of a single higher-order factor was
estimated in a separate function: \code{reliabilityL2}.  It is now available
for multiple higher-order factors in the same model, and from the same
\code{\link[=compRelSEM]{compRelSEM()}} function that estimates reliability for first-order
factors, using the \verb{higher=} argument to name higher-order factors in
the fitted model.
}

\section{\code{runMI()} and the \code{lavaan.mi} class functionality}{

The \code{runMI()} function and support for \code{lavaan.mi-class} objects became
such a large part of semTools that it made sense to move that functionality
to its own package.  The \pkg{lavaan.mi} package is now available for users
to fit \code{lavaan} models to their multiply imputed data.  The new package
already fixes many bugs and provides many new features that make the old
semTools \code{lavaan.mi-class} obsolete, so the latter are now defunct.
}

\keyword{internal}
