% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel2lavaan.R, R/semTools-deprecated.R,
%   R/standardizeMx.R
\name{semTools-deprecated}
\alias{lisrel2lavaan}
\alias{semTools-deprecated}
\alias{standardizeMx}
\title{Deprecated functions in package \pkg{semTools}.}
\usage{
lisrel2lavaan(filename = NULL, analyze = TRUE, silent = FALSE, ...)

standardizeMx(object, free = TRUE)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("semTools-deprecated")}.
}
\section{\code{lisrel2lavaan}}{

The \code{lisrel2lavaan} function will no longer be supported, nor will
there be a replacement function.
}

\section{\code{standardizeMx}}{

The \code{standardizeMx} function will no longer be supported, nor will
there be a replacement function. Its functionality is now available in the
\pkg{OpenMx} package, making this function obsolete, with the exception
that \code{\link[OpenMx]{mxStandardizeRAMpaths}} currently only provides
standardized estimates of covariance-structure parameters, whereas
\code{standardizeMx} also provides standardized means.
}

\keyword{internal}
