% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_semTools-deprecated.R, R/longInvariance.R,
%   R/measurementInvariance.R, R/measurementInvarianceCat.R
\name{semTools-deprecated}
\alias{semTools-deprecated}
\alias{longInvariance}
\alias{measurementInvariance}
\alias{measurementInvarianceCat}
\title{Deprecated functions in package \pkg{semTools}.}
\usage{
longInvariance(model, varList, auto = "all", constrainAuto = FALSE,
  fixed.x = TRUE, std.lv = FALSE, group = NULL, group.equal = "",
  group.partial = "", strict = FALSE, warn = TRUE, debug = FALSE,
  quiet = FALSE, fit.measures = "default", baseline.model = NULL,
  method = "satorra.bentler.2001", ...)

measurementInvariance(..., std.lv = FALSE, strict = FALSE, quiet = FALSE,
  fit.measures = "default", baseline.model = NULL,
  method = "satorra.bentler.2001")

measurementInvarianceCat(..., std.lv = FALSE, strict = FALSE,
  quiet = FALSE, fit.measures = "default", baseline.model = NULL,
  method = "default")
}
\description{
The functions listed below are deprecated and will be defunct
  in the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("semTools-deprecated")}.
}
\section{Previous measurement-invariance functions}{

The \code{measurementInvariance}, \code{measurementInvarianceCat}, and
\code{longInvariance} functions will no longer be supported. Instead, use
the \code{\link{measEq.syntax}} function, which is much more flexible and
supports a wider range of data (e.g., any mixture of \code{numeric} and
\code{ordered} indicators, any combination of multiple groups and repeated
measures, models fit to multiple imputations with \code{\link{runMI}}).
}

\keyword{internal}
